Cleaning up style mistakes

+++ Source/FirstParty/Src/model/Level_LoadingPart.cpp
@@ -437,7 +437,7 @@ float computeArea(const std::vector<b2Vec2>& vertices)
 {
     auto area = 0.0f;
     b2Vec2 origin(0.0f, 0.0f);
-    for(int32 i = 0; i < vertices.size(); ++i)
+    for(size_t i = 0; i < vertices.size(); ++i)
     {
         b2Vec2 d1 = vertices[i] - origin;
         b2Vec2 d2 = vertices[(i + 1) % vertices.size()] - origin;
@@ -493,7 +493,7 @@ void Level::parsePhysics(tinyxml2::XMLElement* physic,
             float area = computeArea(vertices);
             // If the area is negative, the polygon is either messed up or defined in wrong order.
             // A wrong order can be fixed by simply reversing all given vertices.
-            if (area < 0)
+            if(area < 0)
                 std::reverse(begin(vertices), end(vertices));
             ps->Set(vertices.data(), vertices.size());
             shapes.push_back(std::move(ps));
